카카오로그인

+++ src/main/java/com/phl/cocolo/dto/MemberUpdateDTO.java
@@ -16,7 +16,7 @@ public class MemberUpdateDTO {
     private String memberPhone;
     private MultipartFile memberProfile;
     private String memberProfileName;
-    private int memberPoint;
+    private long memberPoint;
     private String memberInterest;
     private String memberNickName;
     private String memberLevel;
